added npc creation

+++ bot/commands/Dungeons&Dragons/nonPlayableCharacterCreation.js
@@ -1,48 +1,42 @@
 const Player = require('../../../database/models/Player');
-const PlayerCharacter = require('../../../database/models/PlayerCharacter');
+const NonPlayableCharacter = require('../../../database/models/NonPlayableCharacter');
 const { MessageEmbed, MessageActionRow, MessageButton, MessageSelectMenu } = require('discord.js');
 
-const QUESTIONS_ARRAY = require('../../jsonDb/characterCreationQuestions.json');
+const QUESTIONS_ARRAY = require('../../jsonDb/npcCreationQuestions.json');
 const { getCharacterEmbed, getCharacterLevelImage } = require('../../otherFunctions/characterEmbed')
 
 module.exports.run = async (bot, message, args) => {
     const characterCreateCategory = message.guild.channels.cache.find(c => c.name == "--CHARACTER CREATION--" && c.type == "GUILD_CATEGORY")
-    let foundPlayer = await Player.findOne({ where: { player_id: message.author.id, server_id: message.guild.id } })
-    if (!foundPlayer) {
-        await Player.create({
-            player_id: message.author.id,
-            player_name: message.author.username,
-            server_id: message.guild.id
-        });
-    } else {
-        await PlayerCharacter.findOne({ where: { player_id: message.author.id, server_id: message.guild.id, status: "CREATING" } })
-        .then((character) => {
-            let name = message.author.username + "-" + message.author.discriminator;
-            let tmpchannel = message.guild.channels.cache.find(channel => channel.name == name.toLowerCase());
-            if (!tmpchannel){if (character) {tmpchannel = message.guild.channels.cache.find(channel => channel.name == character.get("name"));}}
-            if(tmpchannel){tmpchannel.delete();}
-            if (character) { character.destroy();} });
-    }
-    let newCharacter = await PlayerCharacter.create({
-        player_id: message.author.id,
+    if (!message.member.roles.cache.has(message.guild.roles.cache.find(role => role.name.includes('Dungeon Master')).id)) return message.channel.send({ content: 'You\'re not a dm, get lost kid!' }).then(msg => { setTimeout(() => msg.delete(), 5000) }).catch(err => console.log(err));
+    
+    await NonPlayableCharacter.findOne({ where: { creator_id: message.author.id, server_id: message.guild.id, status: "CREATING" } }).then((character) => { 
+        let name = message.author.username + "-" + message.author.discriminator;
+        let tmpchannel = message.guild.channels.cache.find(channel => channel.name == name.toLowerCase());
+        if (!tmpchannel){
+            if (character) {
+                tmpchannel = message.guild.channels.cache.find(channel => channel.name == character.get("name"));
+            }
+        }
+        if(tmpchannel){tmpchannel.delete();}
+
+        if (character) {
+             character.destroy() 
+            } 
+    });
+    
+    let newCharacter = await NonPlayableCharacter.create({
+        creator_id: message.author.id,
         server_id: message.guild.id,
         status: "CREATING"
     });
-
     if (!characterCreateCategory) {
-    return message.channel.send({ content: 'There is no category named \"--CHARACTER CREATION--\"!' })
-        .then(msg => { setTimeout(() => msg.delete(), 3000) })
-        .catch(err => console.log(err));
+        return message.channel.send({ content: 'There is no category named \"--CHARACTER CREATION--\"!' })
+            .then(msg => { setTimeout(() => msg.delete(), 3000) })
+            .catch(err => console.log(err));
     }
     message.guild.channels.cache.forEach(channel => {
         if (channel.name == `${message.author.username.toLowerCase()}-${message.author.discriminator}`)  return message.channel.send({ content: 'You already created a channel before!' }).then(msg => { setTimeout(() => msg.delete(), 3000) }).catch(err => console.log(err));
     });
-
-    if (message.member.roles.cache.has(message.guild.roles.cache.find(role => role.name.includes('Dungeon Master')).id)) {
-        return message.channel.send({ content: 'You\'re not a player, get lost kid!' })
-            .then(msg => { setTimeout(() => msg.delete(), 5000) })
-            .catch(err => console.log(err));
-    }
     message.guild.channels.create(`${message.author.username}-${message.author.discriminator}`, "text").then(async createdChannel => {
         createdChannel.setParent(characterCreateCategory, { lockPermission: false });
         createdChannel.permissionOverwrites.set([{ id: message.author, allow: ['VIEW_CHANNEL'] }, { id: message.guild.roles.cache.find(role => role.name.includes('Dungeon Master')), allow: ['VIEW_CHANNEL'] }, { id: message.channel.guild.roles.everyone, deny: ['VIEW_CHANNEL'] }]);
@@ -50,14 +44,15 @@ module.exports.run = async (bot, message, args) => {
             for (let index = 0; index < QUESTIONS_ARRAY.length; index++) {
                 await characterCreationQuestion(QUESTIONS_ARRAY[index], createdChannel, newCharacter, message, bot, index)
             }
+
             const messageComponents = new MessageActionRow()
                 .addComponents(
                     new MessageButton()
-                        .setCustomId('approve-character-button')
+                        .setCustomId('approve-npc-button')
                         .setLabel('Approve')
                         .setStyle('SUCCESS'),
                     new MessageButton()
-                        .setCustomId('decline-character-button')
+                        .setCustomId('decline-npc-button')
                         .setLabel('Decline')
                         .setStyle('DANGER')
                 );
@@ -69,8 +64,8 @@ module.exports.run = async (bot, message, args) => {
 }
 
 module.exports.help = {
-    name: "createCharacter",
-    alias: ["cc"],
+    name: "createNPC",
+    alias: ["cNPC"],
     description: "Creates a new channel with questions about your new character",
     category: "Dungeons & Dragons"
 }
